+++ src/main/java/br/com/inova/sigin/pessoa/repository/PessoaTipoRepository.java
@@ -4,5 +4,6 @@ import br.com.inova.sigin.pessoa.entity.PessoaTipo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PessoaTipoRepository extends JpaRepository<PessoaTipo, Long> {
+    boolean existsByPessoaIdAndTipoPessoaId(Long pessoaId, Long tipoPessoaId);
 
 }